API: Separate "LBR Region" from actual region
While we use Route53 we must map non-amazon cloud's to the amazon regions,
as LBR only has a fixed list of locations.  This means that the "region"
as far as DNS is concerned is a different thing from the "region"
cloud-wise.

I call the two variables lbr_region and region to disambiguate.