API: Add backup capability
BPB-43

When creating a cluster, there are two new API parameters: "backup_count",
an integer defaulting to 24, which gives the number of backups to keep and
"backup_schedule", either a cron-style schedule or a period in hours at
which a backup will be made.

The data is backed up with mysqldump activated by cron, rotated with
logrotate and uploaded with s3cmd to an s3 bucket. Each cluster has it's
own IAM account.

We now require AMI to have s3cmd installed.

There is still no way for the user to actually see/download her backups.