API: inform django & api consumers about list of backups
BPB-43

For each node a list of backup file is kept in the django app.  Every time
a backup occurs, the node uploads the current list of files, replacing
the list django has stored for that node.

This data is used to present the user links to the s3 objects.