API: retry S3 permissions up to 15 times a 2s intervals
S3 takes some time to notice new IAM users. As we create a new IAM user
and then immediately grant it an S3 permission we must allow for this.
Simply adding a sleep is either unreliable (for short sleep) or potentially
wasteful (long sleep). Instead lets retry until it works, or 30s.