API: only generate SSL keys once
A Model's constructor is called every time the a model is instantiated,
not just when a new model is being created.  This is therefore the wrong
place to generate SSL keys, which should only be done once.

Launch is a better place.  We also manually add some idempotence code to
the generate_keys function to ensure the keys never change after being
generated.