api: remove all migrations
BPB-61 #resolve

Currently every deployment is up to date for migrations, it is therefore
possible for us to remove all the old migrations and start fresh. This
requires the following to be done as a one off:

1. Ensure migrations up to 22 are applied.
2. Check out (at least) this commit
3. Run ./manage.py migrate --fake --delete-ghost-migrations

This fixes BPB-61 as we are using the same AUTH_USER_MODEL right from the
start.