api.models: do cloud provider launch in asyncronous stage
GCE takes some time to launch a node (because we wait for disk creation).
This caused an issue because the syncronous part is run in series, and
we ended up with one node completing launch before the next had been
assigned a nid. This is bad because it fetch cloud_config data and was
given an invalid address for it's neighbour.

This change mitigates this by moving the launch in to the parallel, async
stage.