api.tasks: reload Node and Cluster objects before calling methods
The node and cluster object that is passed into each task is
unchanged from how it existed in controller, even if it has been
save()ed.  This commit reloads the object before it is used so
any save()ed changes will be seen.

Perhaps a better solution would be to pass and return the node
objects along. This is made complicated by the groups used in the
controller.  This will do until we need to worry about
performance.