api.models: attempt salt test.ping before state.highstate
This way if the machine is not responding yet, we error out before sending
state.highstate. It seems that in some cases where we get a error that the
minion didn't answer, the minion does *eventually* get the highstate and
we end up executing it multpile times.

This *may* speed things up.

It would better if we could detect/manage failure/timeout better, but I
can't get to the bottom of the logic inside salt. It may be related to ZMQ
and its habit of hanging on to messages.