api.models: shrink buffer_pool_size
EC2 t1.micro servers were failing to mmap the buffer pool as it was too
large.  The 630MiB that the server has is the physical ram, it only has
590MiB of ram for applications. This is already accounted for in the 30%
free space we try and leave. However the 128MB innodb buffer pool is
enough to swing it. This change accounts for that.