api.views: simplify NodeViewSet.update()
BPB-100

This function because broken with my last commit because free_allowed is
now a read-only property, which this function attempts to set. This commit
fixes this by removing a whole bunch of sets that don't make sense.

The change in flavor is accomplished by setting new_object.flavor... there
is no need to set the individual properties of the flavor.