api.initial_data: remove disk-based gce flavors
We currently attach disks to these anyway, making them identical to their
non "-d" counterpart.  If we improve the gce support to allow instance
storage, we probably should just leave it as the same official flavor and
map the fixed/variable storage difference to the instance type. This would
keep down the length of the flavor list.