api.controller: refresh running nodes before copying backend data
BPB-188

We copy backend data through tinc. In order for the new node to communicate
with the existing nodes they must have it's tinc private key. Therefore
we mush do the salt refresh on the existing nodes before the copy.

However we must salt-refresh the new node after the copy so that mysqld
is started.